backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
imblearn
pydantic
shap
xgboost
joblib
python-multipart
google-generativeai
pymupdf
importlib-metadata
